Extract implementation to hook
