+++ packages/remix-react/browser.tsx
@@ -20,6 +20,7 @@ import {
   getSingleFetchDataStrategy,
 } from "./single-fetch";
 import invariant from "./invariant";
+import { initFogOfWar, useFogOFWarDiscovery } from "./fog-of-war";
 
 /* eslint-disable prefer-let/prefer-let */
 declare global {
@@ -78,20 +79,6 @@ let hmrRouterReadyPromise = new Promise<Router>((resolve) => {
   // in the console. The promise is never rejected.
   return undefined;
 });
-
-type FogOfWarInfo = {
-  controller: AbortController | null;
-  // Currently rendered links that may need prefetching
-  nextPaths: Set<string>;
-  // Paths we know the client can already match, so no need to perform client-side
-  // matching or prefetching for them.  Just an optimization to avoid re-matching
-  // on a larger and larger route tree over time
-  knownGoodPaths: Set<string>;
-  // Routes the server was unable to match - don't ask for them again
-  known404Paths: Set<string>;
-};
-
-let fogOfWar: FogOfWarInfo | null = null;
 
 // @ts-expect-error
 if (import.meta && import.meta.hot) {
@@ -318,17 +305,13 @@ export function RemixBrowser(_props: RemixBrowserProps): ReactElement {
       }
     }
 
-    if (
-      window.__remixContext.future.unstable_fogOfWar === true &&
-      !window.__remixContext.isSpaMode
-    ) {
-      fogOfWar = {
-        controller: null,
-        nextPaths: new Set<string>(),
-        knownGoodPaths: new Set<string>(),
-        known404Paths: new Set<string>(),
-      };
-    }
+    let { enabled: isFogOfWarEnabled, patchRoutesOnMiss } = initFogOfWar(
+      window.__remixManifest,
+      window.__remixRouteModules,
+      window.__remixContext.future,
+      window.__remixContext.isSpaMode,
+      window.__remixContext.basename
+    );
 
     // We don't use createBrowserRouter here because we need fine-grained control
     // over initialization to support synchronous `clientLoader` flows.
@@ -354,19 +337,8 @@ export function RemixBrowser(_props: RemixBrowserProps): ReactElement {
             window.__remixRouteModules
           )
         : undefined,
-      ...(fogOfWar
-        ? {
-            async unstable_patchRoutesOnMiss({ path, patch }) {
-              if (fogOfWar!.known404Paths.has(path)) return;
-              await fetchAndApplyManifestPatches(
-                [path],
-                fogOfWar!.known404Paths,
-                window.__remixContext.basename,
-                window.__remixManifest.version,
-                patch
-              );
-            },
-          }
+      ...(isFogOfWarEnabled
+        ? { unstable_patchRoutesOnMiss: patchRoutesOnMiss }
         : {}),
     });
 
@@ -426,89 +398,13 @@ export function RemixBrowser(_props: RemixBrowserProps): ReactElement {
   }, [location]);
 
   // eslint-disable-next-line react-hooks/rules-of-hooks
-  React.useEffect(() => {
-    // Don't perform prefetching without flag or if the user has `saveData` enabled
-    if (!fogOfWar || navigator.connection?.saveData === true) {
-      return;
-    }
-
-    // Register a link href for patching
-    function registerPath(path: string | null) {
-      let { knownGoodPaths, known404Paths, nextPaths } = fogOfWar!;
-      if (path && !knownGoodPaths.has(path) && !known404Paths.has(path)) {
-        nextPaths.add(path);
-      }
-    }
-
-    // Fetch patches for all currently rendered links
-    async function fetchPatches() {
-      fogOfWar?.controller?.abort();
-
-      let lazyPaths = getFogOfWarPaths(
-        fogOfWar!,
-        window.__remixContext.basename
-      );
-      if (lazyPaths.length === 0) {
-        return;
-      }
-
-      try {
-        fogOfWar!.controller = new AbortController();
-        await fetchAndApplyManifestPatches(
-          lazyPaths,
-          fogOfWar!.known404Paths,
-          window.__remixContext.basename,
-          window.__remixManifest.version,
-          router.patchRoutes,
-          fogOfWar!.controller.signal
-        );
-      } catch (e) {
-        console.error("Failed to fetch manifest patches", e);
-      }
-    }
-
-    // Register and fetch patches for all initially-rendered links
-    document.body
-      .querySelectorAll("a[data-discover]")
-      .forEach((a) => registerPath(a.getAttribute("href")));
-
-    fetchPatches();
-
-    // Setup a MutationObserver to fetch all subsequently rendered links
-    let debouncedFetchPatches = debounce(fetchPatches, 100);
-
-    function isElement(node: Node): node is Element {
-      return node.nodeType === Node.ELEMENT_NODE;
-    }
-
-    let observer = new MutationObserver((records) => {
-      records.forEach((r) => {
-        [r.target, ...r.addedNodes].forEach((node) => {
-          if (!isElement(node)) return;
-          if (node.tagName === "A" && node.getAttribute("data-discover")) {
-            registerPath(node.getAttribute("href"));
-          } else if (node.tagName !== "A") {
-            node
-              .querySelectorAll("a[data-discover]")
-              .forEach((el) => registerPath(el.getAttribute("href")));
-          }
-          debouncedFetchPatches();
-        });
-      });
-    });
-
-    observer.observe(document.documentElement, {
-      subtree: true,
-      childList: true,
-      attributes: true,
-      attributeFilter: ["data-discover", "href"],
-    });
-
-    return () => {
-      fogOfWar?.controller?.abort();
-      observer.disconnect();
-    };
-  }, []);
+  useFogOFWarDiscovery(
+    router,
+    window.__remixManifest,
+    window.__remixRouteModules,
+    window.__remixContext.future,
+    window.__remixContext.isSpaMode
+  );
 
   // We need to include a wrapper RemixErrorBoundary here in case the root error
   // boundary also throws and we need to bubble up outside of the router entirely.
@@ -542,97 +438,4 @@ export function RemixBrowser(_props: RemixBrowserProps): ReactElement {
       {window.__remixContext.future.unstable_singleFetch ? <></> : null}
     </>
   );
-}
-
-function getFogOfWarPaths(
-  fogOfWar: FogOfWarInfo,
-  basename: string | undefined
-) {
-  let { knownGoodPaths, known404Paths, nextPaths } = fogOfWar;
-  return Array.from(nextPaths.keys()).filter((path) => {
-    if (knownGoodPaths.has(path)) {
-      nextPaths.delete(path);
-      return false;
-    }
-
-    if (known404Paths.has(path)) {
-      nextPaths.delete(path);
-      return false;
-    }
-
-    let matches = matchRoutes(router.routes, path, basename);
-    if (matches) {
-      knownGoodPaths.add(path);
-      nextPaths.delete(path);
-      return false;
-    }
-
-    return true;
-  });
-}
-
-async function fetchAndApplyManifestPatches(
-  paths: string[],
-  known404Paths: Set<string>,
-  basename: string | undefined,
-  version: string,
-  patchRoutes: Router["patchRoutes"],
-  signal?: AbortSignal
-): Promise<void> {
-  let manifestPath = `${basename ?? "/"}/__manifest`.replace(/\/+/g, "/");
-  let url = new URL(manifestPath, window.location.origin);
-  url.searchParams.set("version", version);
-  paths.forEach((path) => url.searchParams.append("paths", path));
-  let data = (await fetch(url, { signal }).then((res) => res.json())) as {
-    notFoundPaths: string[];
-    patches: AssetsManifest["routes"];
-  };
-
-  // Capture this before we apply the patches to the manifest
-  let knownRoutes = new Set(Object.keys(window.__remixManifest.routes));
-
-  // Patch routes we don't know about yet into the manifest
-  let patches: AssetsManifest["routes"] = Object.values(data.patches).reduce(
-    (acc, route) =>
-      !knownRoutes.has(route.id)
-        ? Object.assign(acc, { [route.id]: route })
-        : acc,
-    {}
-  );
-  Object.assign(window.__remixManifest.routes, patches);
-
-  // Track legit 404s so we don't try to fetch them again
-  data.notFoundPaths.forEach((p: string) => known404Paths.add(p));
-
-  // Identify all parentIds for which we have new children to add and patch
-  // in their new children
-  let parentIds = new Set<string | undefined>();
-  Object.values(patches).forEach((patch) => {
-    if (!patch.parentId || !patches[patch.parentId]) {
-      parentIds.add(patch.parentId);
-    }
-  });
-  parentIds.forEach((parentId) =>
-    patchRoutes(
-      parentId || null,
-      createClientRoutes(
-        patches,
-        window.__remixRouteModules,
-        null,
-        window.__remixContext.future,
-        window.__remixContext.isSpaMode,
-        parentId
-      )
-    )
-  );
-}
-
-// Thanks Josh!
-// https://www.joshwcomeau.com/snippets/javascript/debounce/
-function debounce(callback: (...args: unknown[]) => unknown, wait: number) {
-  let timeoutId: number | undefined;
-  return (...args: unknown[]) => {
-    window.clearTimeout(timeoutId);
-    timeoutId = window.setTimeout(() => callback(...args), wait);
-  };
 }
